IMAp evaluation function implemented and returning the support vectors.

+++ src/Visualization.cpp
@@ -39,19 +39,19 @@ void Visualization< T >::createPosNegTemps(){
     for(i = 0; i < size; i++){
         if(samples->getPoint(i)->y == 1){
             for(j = 0; j < dim-1; j++){
-                pos_file << samples->getPoint(i)->x[j] << " ";
+                pos_file << (double)(samples->getPoint(i)->x[j]) << " ";
             }
-            pos_file << samples->getPoint(i)->x[j] << endl;
+            pos_file << (double)(samples->getPoint(i)->x[j]) << endl;
         }else if(samples->getPoint(i)->y == -1){
             for(j = 0; j < dim-1; j++){
-                neg_file << samples->getPoint(i)->x[j] << " ";
+                neg_file << (double)(samples->getPoint(i)->x[j]) << " ";
             }
-            neg_file << samples->getPoint(i)->x[j] << endl;
+            neg_file << (double)(samples->getPoint(i)->x[j]) << endl;
         }else{
             for(j = 0; j < dim-1; j++){
-                und_file << samples->getPoint(i)->x[j] << " ";
+                und_file << (double)(samples->getPoint(i)->x[j]) << " ";
             }
-            und_file << samples->getPoint(i)->x[j] << endl;
+            und_file << (double)(samples->getPoint(i)->x[j]) << endl;
         }
     }
     pos_file.close();
